ensure static memory is not reachable on program exit
